Renamed GraphCategory file
